システムスペックの記述を整理した
トークテーマの追加・更新・削除についてはtalk_themes_specに、
話す人の追加・更新・削除についてはspeakers_specに記述するように変更。
