+++ spec/system/roulettes_spec.rb
@@ -103,7 +103,7 @@ RSpec.describe 'Roulettes', type: :system do
     expect(session_storage_items2['speaker'].size).to eq 1
   end
 
-  scenario 'user reset roulette', js: true do
+  scenario 'user resets roulette', js: true do
     visit roulette_path(@roulette)
     click_button 'スタート'
     find "[data-rotate-target='resultText']", wait: 10
@@ -114,107 +114,6 @@ RSpec.describe 'Roulettes', type: :system do
     session_storage_items2 = JSON.parse(evaluate_script("sessionStorage.getItem('#{@roulette.id}')"))
     expect(session_storage_items2['talk'].size).to eq 4
     expect(session_storage_items2['speaker'].size).to eq 4
-  end
-
-  scenario 'user add a talk theme', js: true do
-    visit roulette_path(@roulette)
-    expect do
-      click_link 'トークテーマを追加する'
-      fill_in 'talk_theme[theme]', with: 'トークテーマ テスト'
-      click_button '登録'
-      expect(find('#talk_themes_list')).to have_content('トークテーマ テスト')
-    end.to change(@roulette.talk_themes, :count).by(1)
-    expect(find('.theme__labelContainer')).to have_content('トークテーマ テスト')
-  end
-
-  scenario 'user modify a talk theme', js: true do
-    visit roulette_path(@roulette)
-    within '#talk_themes_list' do
-      click_link 'トークテーマ', match: :first
-      fill_in 'talk_theme[theme]', with: 'トークテーマ 修正'
-      click_button '更新'
-    end
-    expect(find('#talk_themes_list')).to have_content('トークテーマ 修正')
-    expect(find('.theme__labelContainer')).to have_content('トークテーマ 修正')
-  end
-
-  scenario 'user delete a talk theme', js: true do
-    visit roulette_path(@roulette)
-    talk_theme = find('#talk_themes_list').find('a', match: :first).text
-    expect do
-      within '#talk_themes_list' do
-        find('img', visible: false, match: :first).click
-        expect(page).to_not have_content(talk_theme)
-        expect(page.all('li').count).to eq 3
-      end
-    end.to change(@roulette.talk_themes, :count).by(-1)
-    expect(find('.theme__labelContainer')).to_not have_content(talk_theme)
-    expect(find('.theme__labelContainer').all('li').count).to eq 3
-  end
-
-  scenario 'it toggles a link to text when 10 themes are registered', js: true do
-    FactoryBot.create_list(:talk_theme, 5, roulette: @roulette)
-    visit roulette_path(@roulette)
-    expect do
-      click_link 'トークテーマを追加する'
-      fill_in 'talk_theme[theme]', with: 'トークテーマ テスト'
-      click_button '登録'
-      expect(find('#talk_themes_list')).to have_content('トークテーマ テスト')
-    end.to change(@roulette.talk_themes, :count).by(1)
-    new_link_element = find("[data-testid='new-talk-theme-link']")
-    expect(new_link_element).to have_content('トークテーマは10個以上追加できません')
-    expect(new_link_element).to_not have_selector('a')
-  end
-
-  scenario 'user add a speaker', js: true do
-    visit roulette_path(@roulette)
-    expect do
-      click_link '話す人を追加する'
-      fill_in 'speaker[name]', with: 'ユーザー テスト'
-      click_button '登録'
-      expect(find('#speakers_list')).to have_content('ユーザー テスト')
-    end.to change(@roulette.speakers, :count).by(1)
-    expect(find('.speaker__labelContainer')).to have_content('ユーザー テスト')
-  end
-
-  scenario 'user modify a speaker', js: true do
-    visit roulette_path(@roulette)
-    within '#speakers_list' do
-      click_link 'ユーザー', match: :first
-      fill_in 'speaker[name]', with: 'ユーザー 修正'
-      click_button '更新'
-    end
-    expect(find('#speakers_list')).to have_content('ユーザー 修正')
-    expect(find('.speaker__labelContainer')).to have_content('ユーザー 修正')
-  end
-
-  scenario 'user delete a speaker', js: true do
-    visit roulette_path(@roulette)
-    speaker = find('#speakers_list').find('a', match: :first).text
-    expect do
-      within '#speakers_list' do
-        expect(page).to have_content(speaker)
-        find('img', visible: false, match: :first).click
-        expect(page).to_not have_content(speaker)
-        expect(page.all('li').count).to eq 3
-      end
-    end.to change(@roulette.speakers, :count).by(-1)
-    expect(find('.speaker__labelContainer')).to_not have_content(speaker)
-    expect(find('.speaker__labelContainer').all('li').count).to eq 3
-  end
-
-  scenario 'it toggles a link to text when 10 speakers are registered', js: true do
-    FactoryBot.create_list(:speaker, 5, roulette: @roulette)
-    visit roulette_path(@roulette)
-    expect do
-      click_link '話す人を追加する'
-      fill_in 'speaker[name]', with: 'ユーザー テスト'
-      click_button '登録'
-      expect(find('#speakers_list')).to have_content('ユーザー テスト')
-    end.to change(@roulette.speakers, :count).by(1)
-    new_link_element = find("[data-testid='new-speaker-link']")
-    expect(new_link_element).to have_content('話す人は10人以上追加できません')
-    expect(new_link_element).to_not have_selector('a')
   end
 
   scenario 'user registers a one-line talk theme', js: true do
